pyproject.toml
[project]
name = "etude-rag2"
version = "0.1.0"
description = "RAG system with Google Drive and Vertex AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Google Cloud & Drive
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "langchain-google-vertexai>=2.0.7",
    # Database
    "psycopg2-binary>=2.9.10",
    "pgvector>=0.3.6",
    # LangChain
    "langchain>=0.3.0",
    "langchain-postgres>=0.0.12",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    # Embeddings & Reranking
    "FlagEmbedding>=1.2.11",
    "transformers>=4.44.0",
    "torch>=2.4.0",
    "sentence-transformers>=3.0.0",
    # API Server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "sse-starlette>=2.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # Streamlit UI
    "streamlit>=1.39.0",
    "httpx>=0.27.0",
    # Evaluation
    "ragas>=0.1.19",
    "openevals>=0.1.0",
    "langsmith>=0.3.13",
    # Utilities
    "python-dotenv>=1.0.1",
    # Authentication
    "authlib>=1.3.0",
    "itsdangerous>=2.1.0",
    # Web Search
    "duckduckgo-search>=7.0.0",
    "python-multipart>=0.0.21",
    "google-cloud-secret-manager>=2.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "FlagEmbedding.*",
    "google.*",
    "googleapiclient.*",
    "langchain.*",
    "langchain_core.*",
    "langchain_postgres.*",
    "langchain_google_vertexai.*",
    "langchain_ollama.*",
    "psycopg2.*",
    "pgvector.*",
    "ragas.*",
    "sentence_transformers.*",
    "streamlit",
    "transformers.*",
]
ignore_missing_imports = true
